tests/wpt/web-platform-tests/websockets/opening-handshake/001.html

<!doctype html>
<title>WebSockets: invalid handshake</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/invalid');
  ws.onclose = t.step_func(function(e) {
    assert_false(e.wasClean);
    ws.onclose = t.step_func(function() {assert_unreached()});
    setTimeout(t.step_func(function() {t.done();}), 50)
  });
   ws.onmessage = ws.onopen = t.step_func(function() {assert_unreached()});
}, null, {timeout:9900});
</script>
