LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/opening-handshake/001-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/invalid' failed: The server did not accept the WebSocket handshake.

PASS WebSockets: invalid handshake
